Fix na assinatura do metodo.
git-svn-id: https://mbsutil.googlecode.com/svn@110 6ba988c5-0224-0410-acf2-7d9395eead91

+++ Bufer.cpp
@@ -155,7 +155,7 @@ int Buffer::writeLong(unsigned long ul)
 	return 0;
 }
 
-int Buffer::readString(char *dest, int max)
+int Buffer::readString(char *dest, unsigned long max)
 {
 	unsigned long tamstr=strlen((char*)pntL)+1;
 	if(disponiveis()<tamstr)
